AI/requirements.txt
opencv-python
numpy
av
cvzone
ultralytics
python-dotenv
psycopg2
psycopg2-binary
pgvector
requests
torchreid
pillow
pandas
rfdetr
matplotlib
supervision